+++ src/main/java/com/emse/spring/faircorp/web/RoomController.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -40,13 +41,13 @@ public class RoomController
 
         if (room == null)
         {
-            room = roomDao.save(new Room(dto.getName(),dto.getLevel()));
+            room = roomDao.save(new Room(dto.getName(), dto.getLevel(), dto.getBuildingId()));
         }
         else
         {
             room.setLevel(dto.getLevel());
             room.setName(dto.getName());
-            room.setLights(dto.getLights());
+            room.setBuildingId(dto.getBuildingId());
             roomDao.save(room);
         }
 
@@ -62,20 +63,32 @@ public class RoomController
     @GetMapping(path = "/{id}/lights")
     public List<LightDto> findAllLights(@PathVariable Long id)
     {
-        return roomDao.findById(id).get().getLights()
-                                         .stream()
-                                         .map(LightDto::new)
-                                         .collect(Collectors.toList());
+        ArrayList<Light> list = new ArrayList<Light>();
+
+        for (Light light : lightDao.findAll())
+        {
+            if (light.getRoom() == id)
+            {
+                list.add(light);
+            }
+        }
+
+        return list.stream()
+                   .map(LightDto::new)
+                   .collect(Collectors.toList());
     }
 
-    @PutMapping(path = "/{id}/switchLight")
+    @PutMapping(path = "/{id}/switchLights")
     public RoomDto switchStatus(@PathVariable Long id)
     {
         Room room = roomDao.findById(id).orElseThrow(IllegalArgumentException::new);
 
-        for (Light light : room.getLights())
+        for (Light light : lightDao.findAll())
         {
-            light.setStatus(light.getStatus() == Status.ON ? Status.OFF: Status.ON);
+            if (light.getRoom() == id)
+            {
+                light.setStatus(light.getStatus() == Status.ON ? Status.OFF: Status.ON);
+            }
         }
         return new RoomDto(room);
     }
@@ -83,9 +96,12 @@ public class RoomController
     @DeleteMapping(path = "/{id}")
     public void delete(@PathVariable Long id)
     {
-        for (Light light : roomDao.findById(id).orElseThrow(IllegalArgumentException::new).getLights())
+        for (Light light : lightDao.findAll())
         {
-            lightDao.delete(light);
+            if (light.getRoom() == id)
+            {
+                lightDao.delete(light);
+            }
         }
         roomDao.deleteById(id);
     }
